Added create recipe message and change price number to Integer.

+++ main_app/models.py
@@ -65,7 +65,7 @@ class Recipe(models.Model):
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     title = models.CharField(max_length=255)
     time_minutes = models.IntegerField()
-    price = models.DecimalField(max_digits=5, decimal_places=2)
+    price = models.IntegerField()
     link = models.CharField(max_length=255, blank=True)  # link of the recipe if it's stored online
     ingredients = models.ManyToManyField('Ingredient')
     tags = models.ManyToManyField('Tag')
